feat(6b): replace MclClass with cva-based generateClass namespace
Replaces the 750-entry MclClass/static-object system with
class-variance-authority. Each class group is now an individual cva
function (color, typography, spacing, opacity, border-width) exported
for tree-shaking, plus a `generateClass` namespace object as the
primary named export.

All component call sites across every package are migrated from
`generateClass('BGCOLOR', color)` to
`generateClass.bgColorVariant({ color })`. The old default export is
kept as a `generateClassLegacy` shim marked `@deprecated`; it
delegates to the new cva functions and will be removed in the next
major version.

Deleted: MclClass, ValidateMclClass, all static/*.ts class files,
mcl-theme/ and validator/ directories.
