+++ src/components/manguito-theme/lib/header/HeaderVertical.vue
@@ -63,7 +63,7 @@ defineExpose<{
   <header class="relative" :style="componentWidth">
     <nav
       class="py-sm header-desktop hidden overflow-y-scroll overscroll-contain lg:sticky lg:left-0 lg:top-0 lg:block lg:h-[100vh]"
-      :class="[generateClass('BGCOLOR', bgColor)]"
+      :class="[generateClass.bgColorVariant({ color: bgColor })]"
     >
       <div class="flex h-full flex-col justify-between">
         <div>
@@ -97,7 +97,7 @@ defineExpose<{
           <nav
             class="py-lg px-sm fixed inset-0 overflow-y-scroll overscroll-contain"
             v-if="navOpen"
-            :class="generateClass('BGCOLOR', bgColor)"
+            :class="generateClass.bgColorVariant({ color: bgColor })"
           >
             <div v-if="slots['mobile-content']" class="h-full">
               <slot name="mobile-content" :header-close="closeNav" />
