+++ src/components/manguito-theme/lib/modal/Modal.vue
@@ -153,7 +153,7 @@ defineExpose<{
         @click="closeModal"
         aria-hidden="true"
         class="fixed inset-0 z-[100] overflow-y-auto bg-opacity-70 backdrop-blur"
-        :class="generateClass('BGCOLOR', backdropColor)"
+        :class="generateClass.bgColorVariant({ color: backdropColor })"
       ></section>
     </Transition>
     <Transition
@@ -174,20 +174,20 @@ defineExpose<{
           aria-modal="true"
           :aria-labelledby="title ? titleId : undefined"
           v-click-outside="closeModal"
-          :class="[generateClass('BGCOLOR', color), customClass]"
+          :class="[generateClass.bgColorVariant({ color }), customClass]"
           class="relative max-h-[80vh] overflow-y-scroll overscroll-contain md:max-h-[60vh]"
         >
           <div v-if="showHeader" class="sticky top-0">
             <slot name="header" :close="closeModal" :status="toggle">
               <div
                 class="p-xs border-b-light-4 flex items-center justify-between border-b-2"
-                :class="generateClass('BGCOLOR', color)"
+                :class="generateClass.bgColorVariant({ color })"
               >
                 <h3
                   v-if="title"
                   :id="titleId"
                   class="h3-md"
-                  :class="generateClass('TEXTCOLOR', titleColor)"
+                  :class="generateClass.textColorVariant({ color: titleColor })"
                 >
                   {{ title }}
                 </h3>
@@ -198,7 +198,7 @@ defineExpose<{
                     viewBox="0 0 384 512"
                     aria-hidden="true"
                     class="h-sm transition-opacity duration-300 ease-in hover:opacity-75 focus:opacity-75"
-                    :class="[generateClass('SVGFILL', titleColor)]"
+                    :class="[generateClass.svgFillColorVariant({ color: titleColor })]"
                   >
                     <!--! Font Awesome Free 6.4.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. -->
                     <path
