+++ src/components/manguito-theme/lib/sidebar/Sidebar.vue
@@ -143,7 +143,7 @@ defineExpose<{
         @click="closeSidebar"
         aria-hidden="true"
         class="fixed inset-0 z-[100] overflow-y-auto bg-opacity-70 backdrop-blur"
-        :class="generateClass('BGCOLOR', backdropColor)"
+        :class="generateClass.bgColorVariant({ color: backdropColor })"
       ></section>
     </Transition>
     <Transition
@@ -164,13 +164,13 @@ defineExpose<{
       >
         <div
           class="relative h-full overflow-y-scroll overscroll-contain"
-          :class="[generateClass('BGCOLOR', color), customClass]"
+          :class="[generateClass.bgColorVariant({ color }), customClass]"
         >
           <div v-if="showHeader" class="sticky top-0 z-10" ref="headerRef">
             <slot name="header" :close="closeSidebar" :status="toggle">
               <div
                 class="p-xs flex items-center justify-between"
-                :class="generateClass('BGCOLOR', color)"
+                :class="generateClass.bgColorVariant({ color })"
               >
                 <button
                   type="button"
@@ -184,7 +184,7 @@ defineExpose<{
                     viewBox="0 0 384 512"
                     aria-hidden="true"
                     class="h-sm transition-opacity duration-300 ease-in hover:opacity-75 focus:opacity-75"
-                    :class="[generateClass('SVGFILL', titleColor)]"
+                    :class="[generateClass.svgFillColorVariant({ color: titleColor })]"
                   >
                     <!--! Font Awesome Free 6.4.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. -->
                     <path
@@ -196,7 +196,7 @@ defineExpose<{
                   v-if="title"
                   :id="titleId"
                   class="h3-md"
-                  :class="generateClass('TEXTCOLOR', titleColor)"
+                  :class="generateClass.textColorVariant({ color: titleColor })"
                 >
                   {{ title }}
                 </h3>
@@ -212,7 +212,7 @@ defineExpose<{
                     viewBox="0 0 384 512"
                     aria-hidden="true"
                     class="h-sm transition-opacity duration-300 ease-in hover:opacity-75 focus:opacity-75"
-                    :class="[generateClass('SVGFILL', titleColor)]"
+                    :class="[generateClass.svgFillColorVariant({ color: titleColor })]"
                   >
                     <!--! Font Awesome Free 6.4.2 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. -->
                     <path
