+++ src/components/manguito-theme/lib/theme/__tests__/generateClass.test.ts
@@ -1,82 +1,113 @@
 import { describe, expect, it } from 'vitest'
-import { generateClass } from '../index'
+import {
+  bgColorVariant,
+  borderWidthVariant,
+  bodyTextVariant,
+  fontWeightVariant,
+  generateClass,
+  h1Variant,
+  h2Variant,
+  marginVariant,
+  opacityVariant,
+  outlineColorVariant,
+  focusOutlineColorVariant,
+  paddingVariant,
+  gapVariant,
+  textColorVariant,
+  hoverBgColorVariant,
+} from '../index'
 
-describe('generateClass — color types', () => {
-  it('BGCOLOR: valid color returns mcl-bg-* class', () => {
-    expect(generateClass('BGCOLOR', 'primary')).toBe('mcl-bg-primary')
+describe('generateClass namespace — color variants', () => {
+  it('bgColorVariant: primary returns mcl-bg-primary', () => {
+    expect(bgColorVariant({ color: 'primary' })).toBe('mcl-bg-primary')
+    expect(generateClass.bgColorVariant({ color: 'primary' })).toBe('mcl-bg-primary')
   })
-  it('BGCOLOR: invalid color returns single space', () => {
-    expect(generateClass('BGCOLOR', 'neon-green' as any)).toBe(' ')
+  it('bgColorVariant: unknown color returns empty string', () => {
+    expect(bgColorVariant({ color: 'neon-green' as any })).toBe('')
   })
-  it('HVBGCOLOR: valid color returns hover:bg-* class', () => {
-    expect(generateClass('HVBGCOLOR', 'secondary')).toBe('hover:bg-secondary')
+  it('hoverBgColorVariant: secondary returns hover:bg-secondary', () => {
+    expect(hoverBgColorVariant({ color: 'secondary' })).toBe('hover:bg-secondary')
+    expect(generateClass.hoverBgColorVariant({ color: 'secondary' })).toBe('hover:bg-secondary')
   })
-  it('TEXTCOLOR: valid color returns text-* class', () => {
-    expect(generateClass('TEXTCOLOR', 'success')).toBe('text-success')
+  it('textColorVariant: success returns text-success', () => {
+    expect(textColorVariant({ color: 'success' })).toBe('text-success')
+    expect(generateClass.textColorVariant({ color: 'success' })).toBe('text-success')
   })
-  it('OUTLINE: valid color returns mcl-outline-* class', () => {
-    expect(generateClass('OUTLINE', 'danger')).toBe('mcl-outline-danger')
+  it('outlineColorVariant: danger returns mcl-outline-danger', () => {
+    expect(outlineColorVariant({ color: 'danger' })).toBe('mcl-outline-danger')
+    expect(generateClass.outlineColorVariant({ color: 'danger' })).toBe('mcl-outline-danger')
   })
-  it('FOCUSOUTLINE: valid color returns focus:mcl-outline-* class', () => {
-    expect(generateClass('FOCUSOUTLINE', 'primary')).toBe(
-      'focus:mcl-outline-primary'
-    )
+  it('focusOutlineColorVariant: primary returns focus:mcl-outline-primary', () => {
+    expect(focusOutlineColorVariant({ color: 'primary' })).toBe('focus:mcl-outline-primary')
+    expect(generateClass.focusOutlineColorVariant({ color: 'primary' })).toBe('focus:mcl-outline-primary')
   })
 })
 
-describe('generateClass — heading types', () => {
-  it('H1: valid size returns h1-* class', () => {
-    expect(generateClass('H1', 'sm')).toBe('h1-sm')
+describe('generateClass namespace — heading variants', () => {
+  it('h1Variant: sm returns h1-sm', () => {
+    expect(h1Variant({ size: 'sm' })).toBe('h1-sm')
+    expect(generateClass.h1Variant({ size: 'sm' })).toBe('h1-sm')
   })
-  it('H2: valid size returns h2-* class', () => {
-    expect(generateClass('H2', 'md')).toBe('h2-md')
+  it('h2Variant: md returns h2-md', () => {
+    expect(h2Variant({ size: 'md' })).toBe('h2-md')
+    expect(generateClass.h2Variant({ size: 'md' })).toBe('h2-md')
   })
-  it('H1: invalid size returns single space', () => {
-    expect(generateClass('H1', 'jumbo' as any)).toBe(' ')
-  })
-})
-
-describe('generateClass — text types', () => {
-  it('BODYTEXT: valid size returns text-* class', () => {
-    expect(generateClass('BODYTEXT', 'md')).toBe('text-md')
-  })
-  it('FONTWEIGHT: bold returns font-bold', () => {
-    expect(generateClass('FONTWEIGHT', 'bold')).toBe('font-bold')
-  })
-  it('TEXTALIGN: center returns text-center', () => {
-    expect(generateClass('TEXTALIGN', 'center')).toBe('text-center')
+  it('h1Variant: unknown size returns empty string', () => {
+    expect(h1Variant({ size: 'jumbo' as any })).toBe('')
   })
 })
 
-describe('generateClass — border width types', () => {
-  it('BORDERW: 1 returns border-[1px]', () => {
-    expect(generateClass('BORDERW', 1)).toBe('border-[1px]')
+describe('generateClass namespace — text variants', () => {
+  it('bodyTextVariant: md returns text-md', () => {
+    expect(bodyTextVariant({ size: 'md' })).toBe('text-md')
+    expect(generateClass.bodyTextVariant({ size: 'md' })).toBe('text-md')
   })
-  it('BORDERW: 0 (out of range) returns single space', () => {
-    expect(generateClass('BORDERW', 0 as any)).toBe(' ')
+  it('fontWeightVariant: bold returns font-bold', () => {
+    expect(fontWeightVariant({ weight: 'bold' })).toBe('font-bold')
+    expect(generateClass.fontWeightVariant({ weight: 'bold' })).toBe('font-bold')
   })
-})
-
-describe('generateClass — opacity types', () => {
-  it('OPACITY: 50 returns opacity-50', () => {
-    expect(generateClass('OPACITY', 50)).toBe('opacity-50')
-  })
-  it('OPACITY: 10 returns opacity-10', () => {
-    expect(generateClass('OPACITY', 10)).toBe('opacity-10')
-  })
-  it('OPACITY: 55 (not a multiple of 10) returns single space', () => {
-    expect(generateClass('OPACITY', 55 as any)).toBe(' ')
+  it('textAlignVariant: center returns text-center', () => {
+    expect(generateClass.textAlignVariant({ align: 'center' })).toBe('text-center')
   })
 })
 
-describe('generateClass — spacing types', () => {
-  it('MARGIN: sm returns m-sm', () => {
-    expect(generateClass('MARGIN', 'sm')).toBe('m-sm')
+describe('generateClass namespace — border width variants', () => {
+  it('borderWidthVariant: 1 returns border-[1px]', () => {
+    expect(borderWidthVariant({ width: 1 })).toBe('border-[1px]')
+    expect(generateClass.borderWidthVariant({ width: 1 })).toBe('border-[1px]')
   })
-  it('PADDING: lg returns p-lg', () => {
-    expect(generateClass('PADDING', 'lg')).toBe('p-lg')
+  it('borderWidthVariant: 12 returns border-[12px]', () => {
+    expect(borderWidthVariant({ width: 12 })).toBe('border-[12px]')
   })
-  it('GAP: 0 returns gap-0', () => {
-    expect(generateClass('GAP', '0')).toBe('gap-0')
+  it('borderWidthVariant: out-of-range returns empty string', () => {
+    expect(borderWidthVariant({ width: 0 as any })).toBe('')
+  })
+})
+
+describe('generateClass namespace — opacity variants', () => {
+  it('opacityVariant: 50 returns opacity-50', () => {
+    expect(opacityVariant({ opacity: 50 })).toBe('opacity-50')
+    expect(generateClass.opacityVariant({ opacity: 50 })).toBe('opacity-50')
+  })
+  it('opacityVariant: 10 returns opacity-10', () => {
+    expect(opacityVariant({ opacity: 10 })).toBe('opacity-10')
+  })
+  it('opacityVariant: invalid value returns empty string', () => {
+    expect(opacityVariant({ opacity: 55 as any })).toBe('')
+  })
+})
+
+describe('generateClass namespace — spacing variants', () => {
+  it('marginVariant: sm returns m-sm', () => {
+    expect(marginVariant({ spacing: 'sm' })).toBe('m-sm')
+    expect(generateClass.marginVariant({ spacing: 'sm' })).toBe('m-sm')
+  })
+  it('paddingVariant: lg returns p-lg', () => {
+    expect(paddingVariant({ spacing: 'lg' })).toBe('p-lg')
+    expect(generateClass.paddingVariant({ spacing: 'lg' })).toBe('p-lg')
+  })
+  it('gapVariant: 0 returns gap-0', () => {
+    expect(gapVariant({ spacing: '0' })).toBe('gap-0')
+    expect(generateClass.gapVariant({ spacing: '0' })).toBe('gap-0')
   })
 })
