--- src/components/manguito-theme/lib/theme/static/colors.ts
@@ -1,542 +1,0 @@
-import { ColorPalette } from './theme.types'
-
-export type ColorClass = {
-  [key in ColorPalette]: string
-}
-
-export class MclColor {
-  // bg color class
-  public bgColor: ColorClass = {
-    primary: 'mcl-bg-primary',
-    secondary: 'mcl-bg-secondary',
-    success: 'mcl-bg-success',
-    danger: 'mcl-bg-danger',
-    info: 'mcl-bg-info',
-    warning: 'mcl-bg-warning',
-    'light-1': 'mcl-bg-light-1',
-    'light-2': 'mcl-bg-light-2',
-    'light-3': 'mcl-bg-light-3',
-    'light-4': 'mcl-bg-light-4',
-    'dark-1': 'mcl-bg-dark-1',
-    'dark-2': 'mcl-bg-dark-2',
-    'dark-3': 'mcl-bg-dark-3',
-    'dark-4': 'mcl-bg-dark-4',
-    black: 'mcl-bg-black',
-    white: 'mcl-bg-white',
-    transparent: 'mcl-bg-transparent',
-  }
-  // hover: bg color class
-  public hoverBgColor: ColorClass = {
-    primary: 'hover:bg-primary',
-    secondary: 'hover:bg-secondary',
-    success: 'hover:bg-success',
-    danger: 'hover:bg-danger',
-    info: 'hover:bg-info',
-    warning: 'hover:bg-warning',
-    'light-1': 'hover:bg-light-1',
-    'light-2': 'hover:bg-light-2',
-    'light-3': 'hover:bg-light-3',
-    'light-4': 'hover:bg-light-4',
-    'dark-1': 'hover:bg-dark-1',
-    'dark-2': 'hover:bg-dark-2',
-    'dark-3': 'hover:bg-dark-3',
-    'dark-4': 'hover:bg-dark-4',
-    black: 'hover:bg-black',
-    white: 'hover:bg-white',
-    transparent: 'hover:bg-transparent',
-  }
-  // focus: bg color class
-  public focusBgColor: ColorClass = {
-    primary: 'focus:bg-primary',
-    secondary: 'focus:bg-secondary',
-    success: 'focus:bg-success',
-    danger: 'focus:bg-danger',
-    info: 'focus:bg-info',
-    warning: 'focus:bg-warning',
-    'light-1': 'focus:bg-light-1',
-    'light-2': 'focus:bg-light-2',
-    'light-3': 'focus:bg-light-3',
-    'light-4': 'focus:bg-light-4',
-    'dark-1': 'focus:bg-dark-1',
-    'dark-2': 'focus:bg-dark-2',
-    'dark-3': 'focus:bg-dark-3',
-    'dark-4': 'focus:bg-dark-4',
-    black: 'focus:bg-black',
-    white: 'focus:bg-white',
-    transparent: 'focus:bg-transparent',
-  }
-  // ::before bg color class
-  public beforeBgColor: ColorClass = {
-    primary: 'before:bg-primary',
-    secondary: 'before:bg-secondary',
-    success: 'before:bg-success',
-    danger: 'before:bg-danger',
-    info: 'before:bg-info',
-    warning: 'before:bg-warning',
-    'light-1': 'before:bg-light-1',
-    'light-2': 'before:bg-light-2',
-    'light-3': 'before:bg-light-3',
-    'light-4': 'before:bg-light-4',
-    'dark-1': 'before:bg-dark-1',
-    'dark-2': 'before:bg-dark-2',
-    'dark-3': 'before:bg-dark-3',
-    'dark-4': 'before:bg-dark-4',
-    black: 'before:bg-black',
-    white: 'before:bg-white',
-    transparent: 'before:bg-transparent',
-  }
-  // ::after bg color class
-  public afterBgColor: ColorClass = {
-    primary: 'after:bg-primary',
-    secondary: 'after:bg-secondary',
-    success: 'after:bg-success',
-    danger: 'after:bg-danger',
-    info: 'after:bg-info',
-    warning: 'after:bg-warning',
-    'light-1': 'after:bg-light-1',
-    'light-2': 'after:bg-light-2',
-    'light-3': 'after:bg-light-3',
-    'light-4': 'after:bg-light-4',
-    'dark-1': 'after:bg-dark-1',
-    'dark-2': 'after:bg-dark-2',
-    'dark-3': 'after:bg-dark-3',
-    'dark-4': 'after:bg-dark-4',
-    black: 'after:bg-black',
-    white: 'after:bg-white',
-    transparent: 'after:bg-transparent',
-  }
-  // active:bg color class
-  public bgActiveColor: ColorClass = {
-    primary: 'active:bg-primary',
-    secondary: 'active:bg-secondary',
-    success: 'active:bg-success',
-    danger: 'active:bg-danger',
-    info: 'active:bg-info',
-    warning: 'active:bg-warning',
-    'light-1': 'active:bg-light-1',
-    'light-2': 'active:bg-light-2',
-    'light-3': 'active:bg-light-3',
-    'light-4': 'active:bg-light-4',
-    'dark-1': 'active:bg-dark-1',
-    'dark-2': 'active:bg-dark-2',
-    'dark-3': 'active:bg-dark-3',
-    'dark-4': 'active:bg-dark-4',
-    black: 'active:bg-black',
-    white: 'active:bg-white',
-    transparent: 'active:bg-transparent',
-  }
-  // text color class
-  public textColor: ColorClass = {
-    primary: 'text-primary',
-    secondary: 'text-secondary',
-    success: 'text-success',
-    danger: 'text-danger',
-    info: 'text-info',
-    warning: 'text-warning',
-    'light-1': 'text-light-1',
-    'light-2': 'text-light-2',
-    'light-3': 'text-light-3',
-    'light-4': 'text-light-4',
-    'dark-1': 'text-dark-1',
-    'dark-2': 'text-dark-2',
-    'dark-3': 'text-dark-3',
-    'dark-4': 'text-dark-4',
-    black: 'text-black',
-    white: 'text-white',
-    transparent: 'text-transparent',
-  }
-  // hover text color class
-  public hoverTextColor: ColorClass = {
-    primary: 'hover:text-primary',
-    secondary: 'hover:text-secondary',
-    success: 'hover:text-success',
-    danger: 'hover:text-danger',
-    info: 'hover:text-info',
-    warning: 'hover:text-warning',
-    'light-1': 'hover:text-light-1',
-    'light-2': 'hover:text-light-2',
-    'light-3': 'hover:text-light-3',
-    'light-4': 'hover:text-light-4',
-    'dark-1': 'hover:text-dark-1',
-    'dark-2': 'hover:text-dark-2',
-    'dark-3': 'hover:text-dark-3',
-    'dark-4': 'hover:text-dark-4',
-    black: 'hover:text-black',
-    white: 'hover:text-white',
-    transparent: 'hover:text-transparent',
-  }
-  // focus text color class
-  public focusTextColor: ColorClass = {
-    primary: 'focus:text-primary',
-    secondary: 'focus:text-secondary',
-    success: 'focus:text-success',
-    danger: 'focus:text-danger',
-    info: 'focus:text-info',
-    warning: 'focus:text-warning',
-    'light-1': 'focus:text-light-1',
-    'light-2': 'focus:text-light-2',
-    'light-3': 'focus:text-light-3',
-    'light-4': 'focus:text-light-4',
-    'dark-1': 'focus:text-dark-1',
-    'dark-2': 'focus:text-dark-2',
-    'dark-3': 'focus:text-dark-3',
-    'dark-4': 'focus:text-dark-4',
-    black: 'focus:text-black',
-    white: 'focus:text-white',
-    transparent: 'focus:text-transparent',
-  }
-  // disabled text color class
-  public disabledTextColor: ColorClass = {
-    primary: 'disabled:text-primary',
-    secondary: 'disabled:text-secondary',
-    success: 'disabled:text-success',
-    danger: 'disabled:text-danger',
-    info: 'disabled:text-info',
-    warning: 'disabled:text-warning',
-    'light-1': 'disabled:text-light-1',
-    'light-2': 'disabled:text-light-2',
-    'light-3': 'disabled:text-light-3',
-    'light-4': 'disabled:text-light-4',
-    'dark-1': 'disabled:text-dark-1',
-    'dark-2': 'disabled:text-dark-2',
-    'dark-3': 'disabled:text-dark-3',
-    'dark-4': 'disabled:text-dark-4',
-    black: 'disabled:text-black',
-    white: 'disabled:text-white',
-    transparent: 'disabled:text-transparent',
-  }
-  // SVG fill color class
-  public svgFillColor: ColorClass = {
-    primary: 'fill-primary',
-    secondary: 'fill-secondary',
-    success: 'fill-success',
-    danger: 'fill-danger',
-    info: 'fill-info',
-    warning: 'fill-warning',
-    'light-1': 'fill-light-1',
-    'light-2': 'fill-light-2',
-    'light-3': 'fill-light-3',
-    'light-4': 'fill-light-4',
-    'dark-1': 'fill-dark-1',
-    'dark-2': 'fill-dark-2',
-    'dark-3': 'fill-dark-3',
-    'dark-4': 'fill-dark-4',
-    black: 'fill-black',
-    white: 'fill-white',
-    transparent: 'fill-transparent',
-  }
-  // ring color class
-  public ringColor: ColorClass = {
-    primary: 'ring-primary',
-    secondary: 'ring-secondary',
-    success: 'ring-success',
-    danger: 'ring-danger',
-    info: 'ring-info',
-    warning: 'ring-warning',
-    'light-1': 'ring-light-1',
-    'light-2': 'ring-light-2',
-    'light-3': 'ring-light-3',
-    'light-4': 'ring-light-4',
-    'dark-1': 'ring-dark-1',
-    'dark-2': 'ring-dark-2',
-    'dark-3': 'ring-dark-3',
-    'dark-4': 'ring-dark-4',
-    black: 'ring-black',
-    white: 'ring-white',
-    transparent: 'ring-transparent',
-  }
-  // focus:ring color class
-  public ringFocusColor: ColorClass = {
-    primary: 'focus:ring-primary',
-    secondary: 'focus:ring-secondary',
-    success: 'focus:ring-success',
-    danger: 'focus:ring-danger',
-    info: 'focus:ring-info',
-    warning: 'focus:ring-warning',
-    'light-1': 'focus:ring-light-1',
-    'light-2': 'focus:ring-light-2',
-    'light-3': 'focus:ring-light-3',
-    'light-4': 'focus:ring-light-4',
-    'dark-1': 'focus:ring-dark-1',
-    'dark-2': 'focus:ring-dark-2',
-    'dark-3': 'focus:ring-dark-3',
-    'dark-4': 'focus:ring-dark-4',
-    black: 'focus:ring-black',
-    white: 'focus:ring-white',
-    transparent: 'focus:ring-transparent',
-  }
-  // active:ring color class
-  public ringActiveColor: ColorClass = {
-    primary: 'active:ring-primary',
-    secondary: 'active:ring-secondary',
-    success: 'active:ring-success',
-    danger: 'active:ring-danger',
-    info: 'active:ring-info',
-    warning: 'active:ring-warning',
-    'light-1': 'active:ring-light-1',
-    'light-2': 'active:ring-light-2',
-    'light-3': 'active:ring-light-3',
-    'light-4': 'active:ring-light-4',
-    'dark-1': 'active:ring-dark-1',
-    'dark-2': 'active:ring-dark-2',
-    'dark-3': 'active:ring-dark-3',
-    'dark-4': 'active:ring-dark-4',
-    black: 'active:ring-black',
-    white: 'active:ring-white',
-    transparent: 'active:ring-transparent',
-  }
-  // ring-offset color class
-  public ringOffsetColor: ColorClass = {
-    primary: 'ring-offset-primary',
-    secondary: 'ring-offset-secondary',
-    success: 'ring-offset-success',
-    danger: 'ring-offset-danger',
-    info: 'ring-offset-info',
-    warning: 'ring-offset-warning',
-    'light-1': 'ring-offset-light-1',
-    'light-2': 'ring-offset-light-2',
-    'light-3': 'ring-offset-light-3',
-    'light-4': 'ring-offset-light-4',
-    'dark-1': 'ring-offset-dark-1',
-    'dark-2': 'ring-offset-dark-2',
-    'dark-3': 'ring-offset-dark-3',
-    'dark-4': 'ring-offset-dark-4',
-    black: 'ring-offset-black',
-    white: 'ring-offset-white',
-    transparent: 'ring-offset-transparent',
-  }
-  // outline class
-  public outlineColor: ColorClass = {
-    primary: 'mcl-outline-primary',
-    secondary: 'mcl-outline-secondary',
-    success: 'mcl-outline-success',
-    danger: 'mcl-outline-danger',
-    info: 'mcl-outline-info',
-    warning: 'mcl-outline-warning',
-    'light-1': 'mcl-outline-light-1',
-    'light-2': 'mcl-outline-light-2',
-    'light-3': 'mcl-outline-light-3',
-    'light-4': 'mcl-outline-light-4',
-    'dark-1': 'mcl-outline-dark-1',
-    'dark-2': 'mcl-outline-dark-2',
-    'dark-3': 'mcl-outline-dark-3',
-    'dark-4': 'mcl-outline-dark-4',
-    black: 'mcl-outline-black',
-    white: 'mcl-outline-white',
-    transparent: 'mcl-outline-transparent',
-  }
-  // focus:outline class
-  public focusOutlineColor: ColorClass = {
-    primary: 'focus:mcl-outline-primary',
-    secondary: 'focus:mcl-outline-secondary',
-    success: 'focus:mcl-outline-success',
-    danger: 'focus:mcl-outline-danger',
-    info: 'focus:mcl-outline-info',
-    warning: 'focus:mcl-outline-warning',
-    'light-1': 'focus:mcl-outline-light-1',
-    'light-2': 'focus:mcl-outline-light-2',
-    'light-3': 'focus:mcl-outline-light-3',
-    'light-4': 'focus:mcl-outline-light-4',
-    'dark-1': 'focus:mcl-outline-dark-1',
-    'dark-2': 'focus:mcl-outline-dark-2',
-    'dark-3': 'focus:mcl-outline-dark-3',
-    'dark-4': 'focus:mcl-outline-dark-4',
-    black: 'focus:mcl-outline-black',
-    white: 'focus:mcl-outline-white',
-    transparent: 'focus:mcl-outline-transparent',
-  }
-  // active:outline class
-  public activeOutlineColor: ColorClass = {
-    primary: 'active:mcl-outline-primary',
-    secondary: 'active:mcl-outline-secondary',
-    success: 'active:mcl-outline-success',
-    danger: 'active:mcl-outline-danger',
-    info: 'active:mcl-outline-info',
-    warning: 'active:mcl-outline-warning',
-    'light-1': 'active:mcl-outline-light-1',
-    'light-2': 'active:mcl-outline-light-2',
-    'light-3': 'active:mcl-outline-light-3',
-    'light-4': 'active:mcl-outline-light-4',
-    'dark-1': 'active:mcl-outline-dark-1',
-    'dark-2': 'active:mcl-outline-dark-2',
-    'dark-3': 'active:mcl-outline-dark-3',
-    'dark-4': 'active:mcl-outline-dark-4',
-    black: 'active:mcl-outline-black',
-    white: 'active:mcl-outline-white',
-    transparent: 'active:mcl-outline-transparent',
-  }
-  // btn color class
-  public btnColor: ColorClass = {
-    primary: 'btn-primary',
-    secondary: 'btn-secondary',
-    success: 'btn-success',
-    danger: 'btn-danger',
-    info: 'btn-info',
-    warning: 'btn-warning',
-    'light-1': 'btn-light-1',
-    'light-2': 'btn-light-2',
-    'light-3': 'btn-light-3',
-    'light-4': 'btn-light-4',
-    'dark-1': 'btn-dark-1',
-    'dark-2': 'btn-dark-2',
-    'dark-3': 'btn-dark-3',
-    'dark-4': 'btn-dark-4',
-    black: 'btn-black',
-    white: 'btn-white',
-    transparent: 'btn-transparent',
-  }
-  // list color class
-  public listColor: ColorClass = {
-    primary: 'mcl-list-primary',
-    secondary: 'mcl-list-secondary',
-    success: 'mcl-list-success',
-    danger: 'mcl-list-danger',
-    info: 'mcl-list-info',
-    warning: 'mcl-list-warning',
-    'light-1': 'mcl-list-light-1',
-    'light-2': 'mcl-list-light-2',
-    'light-3': 'mcl-list-light-3',
-    'light-4': 'mcl-list-light-4',
-    'dark-1': 'mcl-list-dark-1',
-    'dark-2': 'mcl-list-dark-2',
-    'dark-3': 'mcl-list-dark-3',
-    'dark-4': 'mcl-list-dark-4',
-    black: 'mcl-list-black',
-    white: 'mcl-list-white',
-    transparent: 'mcl-list-transparent',
-  }
-  // border color classes
-  public borderColor: ColorClass = {
-    primary: 'border-primary',
-    secondary: 'border-secondary',
-    success: 'border-success',
-    danger: 'border-danger',
-    info: 'border-info',
-    warning: 'border-warning',
-    'light-1': 'border-light-1',
-    'light-2': 'border-light-2',
-    'light-3': 'border-light-3',
-    'light-4': 'border-light-4',
-    'dark-1': 'border-dark-1',
-    'dark-2': 'border-dark-2',
-    'dark-3': 'border-dark-3',
-    'dark-4': 'border-dark-4',
-    black: 'border-black',
-    white: 'border-white',
-    transparent: 'border-transparent',
-  }
-  public borderXColor: ColorClass = {
-    primary: 'border-x-primary',
-    secondary: 'border-x-secondary',
-    success: 'border-x-success',
-    danger: 'border-x-danger',
-    info: 'border-x-info',
-    warning: 'border-x-warning',
-    'light-1': 'border-x-light-1',
-    'light-2': 'border-x-light-2',
-    'light-3': 'border-x-light-3',
-    'light-4': 'border-x-light-4',
-    'dark-1': 'border-x-dark-1',
-    'dark-2': 'border-x-dark-2',
-    'dark-3': 'border-x-dark-3',
-    'dark-4': 'border-x-dark-4',
-    black: 'border-x-black',
-    white: 'border-x-white',
-    transparent: 'border-x-transparent',
-  }
-  public borderYColor: ColorClass = {
-    primary: 'border-y-primary',
-    secondary: 'border-y-secondary',
-    success: 'border-y-success',
-    danger: 'border-y-danger',
-    info: 'border-y-info',
-    warning: 'border-y-warning',
-    'light-1': 'border-y-light-1',
-    'light-2': 'border-y-light-2',
-    'light-3': 'border-y-light-3',
-    'light-4': 'border-y-light-4',
-    'dark-1': 'border-y-dark-1',
-    'dark-2': 'border-y-dark-2',
-    'dark-3': 'border-y-dark-3',
-    'dark-4': 'border-y-dark-4',
-    black: 'border-y-black',
-    white: 'border-y-white',
-    transparent: 'border-y-transparent',
-  }
-  public borderTopColor: ColorClass = {
-    primary: 'border-t-primary',
-    secondary: 'border-t-secondary',
-    success: 'border-t-success',
-    danger: 'border-t-danger',
-    info: 'border-t-info',
-    warning: 'border-t-warning',
-    'light-1': 'border-t-light-1',
-    'light-2': 'border-t-light-2',
-    'light-3': 'border-t-light-3',
-    'light-4': 'border-t-light-4',
-    'dark-1': 'border-t-dark-1',
-    'dark-2': 'border-t-dark-2',
-    'dark-3': 'border-t-dark-3',
-    'dark-4': 'border-t-dark-4',
-    black: 'border-t-black',
-    white: 'border-t-white',
-    transparent: 'border-t-transparent',
-  }
-  public borderBottomColor: ColorClass = {
-    primary: 'border-b-primary',
-    secondary: 'border-b-secondary',
-    success: 'border-b-success',
-    danger: 'border-b-danger',
-    info: 'border-b-info',
-    warning: 'border-b-warning',
-    'light-1': 'border-b-light-1',
-    'light-2': 'border-b-light-2',
-    'light-3': 'border-b-light-3',
-    'light-4': 'border-b-light-4',
-    'dark-1': 'border-b-dark-1',
-    'dark-2': 'border-b-dark-2',
-    'dark-3': 'border-b-dark-3',
-    'dark-4': 'border-b-dark-4',
-    black: 'border-b-black',
-    white: 'border-b-white',
-    transparent: 'border-b-transparent',
-  }
-  public borderLeftColor: ColorClass = {
-    primary: 'border-l-primary',
-    secondary: 'border-l-secondary',
-    success: 'border-l-success',
-    danger: 'border-l-danger',
-    info: 'border-l-info',
-    warning: 'border-l-warning',
-    'light-1': 'border-l-light-1',
-    'light-2': 'border-l-light-2',
-    'light-3': 'border-l-light-3',
-    'light-4': 'border-l-light-4',
-    'dark-1': 'border-l-dark-1',
-    'dark-2': 'border-l-dark-2',
-    'dark-3': 'border-l-dark-3',
-    'dark-4': 'border-l-dark-4',
-    black: 'border-l-black',
-    white: 'border-l-white',
-    transparent: 'border-l-transparent',
-  }
-  public borderRightColor: ColorClass = {
-    primary: 'border-r-primary',
-    secondary: 'border-r-secondary',
-    success: 'border-r-success',
-    danger: 'border-r-danger',
-    info: 'border-r-info',
-    warning: 'border-r-warning',
-    'light-1': 'border-r-light-1',
-    'light-2': 'border-r-light-2',
-    'light-3': 'border-r-light-3',
-    'light-4': 'border-r-light-4',
-    'dark-1': 'border-r-dark-1',
-    'dark-2': 'border-r-dark-2',
-    'dark-3': 'border-r-dark-3',
-    'dark-4': 'border-r-dark-4',
-    black: 'border-r-black',
-    white: 'border-r-white',
-    transparent: 'border-r-transparent',
-  }
-}
