--- src/components/manguito-theme/lib/theme/static/index.ts
@@ -1,7 +1,0 @@
-export * from './body-text'
-export * from './borders'
-export * from './colors'
-export * from './heading-size'
-export * from './opacity'
-export * from './spacing'
-export * from './theme.types'
