--- src/components/manguito-theme/lib/theme/static/opacity.ts
@@ -1,45 +1,0 @@
-import { OpacityRange } from './theme.types'
-
-export type Opacity = {
-  [key in OpacityRange]: string
-}
-
-export class MclOpacityClass {
-  // opacity
-  public backgroundOpacity: Opacity = {
-    10: 'bg-opacity-10',
-    20: 'bg-opacity-20',
-    30: 'bg-opacity-30',
-    40: 'bg-opacity-40',
-    50: 'bg-opacity-50',
-    60: 'bg-opacity-60',
-    70: 'bg-opacity-70',
-    80: 'bg-opacity-80',
-    90: 'bg-opacity-90',
-    100: 'bg-opacity-100',
-  }
-  public opacity: Opacity = {
-    10: 'opacity-10',
-    20: 'opacity-20',
-    30: 'opacity-30',
-    40: 'opacity-40',
-    50: 'opacity-50',
-    60: 'opacity-60',
-    70: 'opacity-70',
-    80: 'opacity-80',
-    90: 'opacity-90',
-    100: 'opacity-100',
-  }
-  public outlineOpacity: Opacity = {
-    10: 'outline-opacity-10',
-    20: 'outline-opacity-20',
-    30: 'outline-opacity-30',
-    40: 'outline-opacity-40',
-    50: 'outline-opacity-50',
-    60: 'outline-opacity-60',
-    70: 'outline-opacity-70',
-    80: 'outline-opacity-80',
-    90: 'outline-opacity-90',
-    100: 'outline-opacity-100',
-  }
-}
