+++ src/components/manguito-theme/package.json
@@ -56,6 +56,7 @@
   "dependencies": {
     "@vueuse/core": "^12.0.0",
     "@vueuse/integrations": "^14.3.0",
+    "class-variance-authority": "^0.7.1",
     "focus-trap": "^8.2.0",
     "isomorphic-dompurify": "^3.12.0",
     "vue": "^3.5.13"
