+++ src/components/mcl-forms/lib/mcl-input-switch/MclInputSwitch.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { ColorPalette } from '@bobbykim/manguito-theme'
-import generateClass from '@bobbykim/manguito-theme'
+import { generateClass } from '@bobbykim/manguito-theme'
 import { computed, ref } from 'vue'
 import type { ColorMap } from '../common/index.types'
 import type { InputSizeType } from '../common/index.types'
@@ -89,8 +89,8 @@ const switchClass = computed<string>(() => {
    */
   const { sliderColor, onColor, offColor, rounded } = props
   const classArray: string[] = [
-    generateClass('BGCOLOR', offColor),
-    generateClass('BEFOREBG', sliderColor),
+    generateClass.bgColorVariant({ color: offColor }),
+    generateClass.beforeBgColorVariant({ color: sliderColor }),
     peerBgColor[onColor],
   ]
   if (rounded) {
