+++ src/components/mcl-header/lib/mcl-header-b/NavCollapseVertical.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { BodyText, ColorPalette } from '@bobbykim/manguito-theme'
-import generateClass, { Collapse } from '@bobbykim/manguito-theme'
+import { generateClass, Collapse } from '@bobbykim/manguito-theme'
 import { computed, ref } from 'vue'
 import type { MenuCollapseType, MenuItemType } from '../common/index.types'
 type NavLocationType = 'desktop' | 'mobile'
@@ -59,24 +59,24 @@ const getNavId = computed<string>(() => {
 
 const colorClass = computed<string>(() => {
   const { textColor, textSize, dHl, hlColor, fontBold } = props
-  const classArray: string[] = [generateClass('BODYTEXT', textSize)]
+  const classArray: string[] = [generateClass.bodyTextVariant({ size: textSize })]
   if (!dHl) {
     const nonHlTextClass: string[] = [
-      generateClass('TEXTCOLOR', textColor),
-      generateClass('HVTEXTCOLOR', hlColor),
-      generateClass('FCTEXTCOLOR', hlColor),
+      generateClass.textColorVariant({ color: textColor }),
+      generateClass.hoverTextColorVariant({ color: hlColor }),
+      generateClass.focusTextColorVariant({ color: hlColor }),
     ]
     classArray.push(nonHlTextClass.join(' '))
   }
   if (dHl) {
-    const highlightColor: string = generateClass('BEFOREBG', hlColor)
+    const highlightColor: string = generateClass.beforeBgColorVariant({ color: hlColor })
     const hlClass: string =
       'before:inset-y-0 before:left-0 before:transition-[width] before:duration-300 before:ease-linear before:w-0 hover:before:w-full focus:before:w-full '
     classArray.push(hlClass + highlightColor)
     const hlTextClass: string[] = [
-      generateClass('TEXTCOLOR', hlColor),
-      generateClass('HVTEXTCOLOR', textColor),
-      generateClass('FCTEXTCOLOR', textColor),
+      generateClass.textColorVariant({ color: hlColor }),
+      generateClass.hoverTextColorVariant({ color: textColor }),
+      generateClass.focusTextColorVariant({ color: textColor }),
     ]
     classArray.push(hlTextClass.join(' '))
   }
@@ -95,7 +95,7 @@ const collapseHighlightClass = computed<string>(() => {
     'before:absolute before:w-xs before:h-full before:mcl-bg-primary before:bg-opacity-70 '
   const hlLocation: string =
     navLocation === 'desktop' ? 'before:left-0 ' : 'before:right-0 '
-  classArray.push(classNames + hlLocation + generateClass('BEFOREBG', hlColor))
+  classArray.push(classNames + hlLocation + generateClass.beforeBgColorVariant({ color: hlColor }))
   return classArray.join(' ')
 })
 const childItemColorClass = computed<string>(() => {
@@ -109,22 +109,22 @@ const childItemColorClass = computed<string>(() => {
           ? 'sm'
           : 'xs'
   const classArray: string[] = [
-    generateClass('TEXTCOLOR', textColor),
-    generateClass('BODYTEXT', childBodyText),
+    generateClass.textColorVariant({ color: textColor }),
+    generateClass.bodyTextVariant({ size: childBodyText }),
   ]
   if (navLocation === 'mobile') {
     classArray.push('text-end')
   }
   if (!dHl) {
     const nonHlTextClass: string[] = [
-      generateClass('TEXTCOLOR', textColor),
-      generateClass('HVTEXTCOLOR', hlColor),
-      generateClass('FCTEXTCOLOR', hlColor),
+      generateClass.textColorVariant({ color: textColor }),
+      generateClass.hoverTextColorVariant({ color: hlColor }),
+      generateClass.focusTextColorVariant({ color: hlColor }),
     ]
     classArray.push(nonHlTextClass.join(' '))
   }
   if (dHl) {
-    const highlightColor: string = generateClass('BEFOREBG', hlColor)
+    const highlightColor: string = generateClass.beforeBgColorVariant({ color: hlColor })
     const hlClass: string =
       'before:inset-y-0 before:duration-300 before:ease-linear before:w-0 hover:before:w-full focus:before:w-full before:bg-opacity-70 '
     const hlLocation: string =
@@ -133,9 +133,9 @@ const childItemColorClass = computed<string>(() => {
         : 'before:left-full hover:before:left-0 focus:before:left-0 before:transition-[all] '
     classArray.push(hlClass + hlLocation + highlightColor)
     const hlTextClass: string[] = [
-      generateClass('TEXTCOLOR', hlColor),
-      generateClass('HVTEXTCOLOR', textColor),
-      generateClass('FCTEXTCOLOR', textColor),
+      generateClass.textColorVariant({ color: hlColor }),
+      generateClass.hoverTextColorVariant({ color: textColor }),
+      generateClass.focusTextColorVariant({ color: textColor }),
     ]
     classArray.push(hlTextClass.join(' '))
   }
